ci: isolate the persistent-tasks lane, split the GHQ file (round 5)
Round-4 wall clock 10m53, set by two lanes:

- aqua 9m41: JULIA_NUM_PRECOMPILE_TASKS=2 did not stop the persistent-tasks
  attempt-1 subprocess death, so the check still runs twice (~4 min). It now
  lives in its own file/lane (aqua_persistent_tasks_tests.jl, group 27) paired
  with uq_edge_cases; aqua_tests keeps the rest of test_all at full strength.
- ghquadrature 7m11 as a lane opener: split at the family-cases boundary into
  part 1 (nodes/kernel/main/NPF/UQ/dedup) and part 2 (RE families, anisotropic,
  refinement, #98/#151 oracles) + copulas in group 26.

14 jobs; every lane now carries <= ~5-6 min of measured content.
