+++ .github/workflows/CI.yml
@@ -34,8 +34,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        group: ["8,25", "17,21", "18,22", "2,24", "19", "13,14", "6,11",
-                "7,23", "5,16", "4,20", "9,10", "3,12", "1,15"]
+        group: ["8,25", "17,21", "18,26", "2,24", "19", "22,27", "13,14",
+                "6,11", "7,23", "5,16", "4,20", "9,10", "3,12", "1,15"]
     steps:
       - uses: actions/checkout@v7
       - uses: julia-actions/setup-julia@v3
