+++ .github/workflows/Coverage.yml
@@ -31,7 +31,7 @@ jobs:
       # batches are bigger and two at once would strain the runner's RAM).
       matrix:
         group: ["1,2", "3,4", "5,6", "7,8", "9,10", "11,12",
-                "13,14", "15,16", "17,18", "19,20", "21,22", "23,24,25"]
+                "13,14", "15,16", "17,18", "19,20", "21,22", "23,24,25", "26,27"]
     steps:
       - uses: actions/checkout@v7
       - uses: julia-actions/setup-julia@v3
